Renamed module directory from "ext" to "GNUpod"
All the modules in ext/ are declaring to be package
GNUpod::something. So lets do the right thing and
put them in subdirectory GNUpod. Together with the
patch that put the default perl location in all the
gnupod tools' #! lines, this has the added benefit
that you can run gnupod tools from the src directory
without building or installing them.